+++ Pandora/Paradigm/Inventory.hs
@@ -2,5 +2,5 @@ module Pandora.Paradigm.Inventory (module Exports) where
 
 import Pandora.Paradigm.Inventory.Optics as Exports
 import Pandora.Paradigm.Inventory.Storage as Exports
-import Pandora.Paradigm.Inventory.Stateful as Exports
+import Pandora.Paradigm.Inventory.State as Exports
 import Pandora.Paradigm.Inventory.Environmental as Exports