+++ Pandora/Paradigm/Inventory/Optics.hs
@@ -3,15 +3,15 @@ module Pandora.Paradigm.Inventory.Optics
 
 import Pandora.Core.Morphism ((.), (!))
 import Pandora.Paradigm.Basis.Product (Product ((:*:)))
-import Pandora.Paradigm.Inventory.Stateful (Stateful (Stateful), statefully)
+import Pandora.Paradigm.Inventory.State (State (State), statefully)
 import Pandora.Paradigm.Inventory.Storage (Storage (Storage), access, position, retrofit)
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$)))
 import Pandora.Pattern.Functor.Extractable (Extractable (extract))
 import Pandora.Pattern.Functor.Adjoint (Adjoint ((-|), (|-)))
 import Pandora.Pattern.Functor.Divariant (($))
 
-instance Adjoint (Storage s) (Stateful s) where
-	v -| f = Stateful $ \s -> (:*:) s . f . Storage $ s :*: (v !)
+instance Adjoint (Storage s) (State s) where
+	v -| f = State $ \s -> (:*:) s . f . Storage $ s :*: (v !)
 	Storage (s :*: f) |- g = extract . statefully s . g $ f s
 
 infixr 0 :-.