+++ CHANGELOG.md
@@ -131,3 +131,5 @@
 * Define `Adaptable` type class for fitting effects in schemas
 * Define `Failable` and `failure` for adaptable `Conclusion`
 * Define `Optional` and `nothing` for adaptable `Maybe`
+* Rename `Stateful` to `State`, `get` to `current`
+* Define `Stateful` adaptable constraint and adapt `get`, `modify` and `put`